Fixed Admin Dashboard Trouble
Instead of handleLogin(e) defaulting to `${API_BASE}/auth/login`, an endpoint is created to be either the default or the admin login. The endpoint is determined by a check as to whether a given login email is the admin email. (If the password is wrong then it still keeps you out)